test(structure): 테스트 디렉토리 구조와 공통 지원 코드 정리
- 테스트 타깃을 기능과 계층 기준 디렉토리로 재배치함
- 공통 fixture, spy, helper를 TestSupport로 분리함
- 현재 앱 코드와 맞지 않는 레거시 disabled 테스트를 정리함

Refs: #54
