+++ NailClient/NailClientTests/Features/AIGeneration/AINailGenerationDetailItemTests.swift
