+++ NailClient/NailClientTests/Features/AIGeneration/AINailGenerationPushOpenTests.swift
@@ -10,11 +10,9 @@ struct AINailGenerationPushOpenTests {
         let parentJobId = UUID(uuidString: "bbbbbbbb-bbbb-4bbb-8bbb-bbbbbbbbbbbb")!
         let resultImageURL = "https://example.com/push-result.png"
         let service = PushOpenAINailGenerationServiceSpy(
-            statusResponse: NailGenJobStatusResponse(
+            statusResponse: NailGenerationTestFixtures.makeStatusResponse(
                 status: .completed,
                 resultImageURL: resultImageURL,
-                errorCode: nil,
-                errorMessage: nil,
                 parentJobId: parentJobId.uuidString,
                 refinementTurn: 2,
                 canRefine: true
@@ -42,11 +40,9 @@ struct AINailGenerationPushOpenTests {
         let createdAt = Date(timeIntervalSince1970: 1_736_000_123)
         let resultImageURL = "https://example.com/push-opened.png"
         let service = PushOpenAINailGenerationServiceSpy(
-            statusResponse: NailGenJobStatusResponse(
+            statusResponse: NailGenerationTestFixtures.makeStatusResponse(
                 status: .completed,
                 resultImageURL: resultImageURL,
-                errorCode: nil,
-                errorMessage: nil,
                 parentJobId: parentJobId.uuidString,
                 refinementTurn: 1,
                 canRefine: false
@@ -69,69 +65,5 @@ struct AINailGenerationPushOpenTests {
         #expect(item.parentJobId == parentJobId)
         #expect(item.refinementTurn == 1)
         #expect(item.isLiked == false)
-    }
-}
-
-@MainActor
-private final class PushOpenAINailGenerationServiceSpy: AINailGenerationServicing {
-    private let statusResponse: NailGenJobStatusResponse
-    private(set) var statusCallCount: Int = 0
-
-    init(statusResponse: NailGenJobStatusResponse) {
-        self.statusResponse = statusResponse
-    }
-
-    func issueNailGenerationUploadURL(
-        kind: NailGenUploadKind,
-        ext: String,
-        contentType: String,
-        bytes: Int,
-        jobId: UUID?
-    ) async throws -> NailGenUploadURLResponse {
-        _ = kind
-        _ = ext
-        _ = contentType
-        _ = bytes
-        return NailGenUploadURLResponse(
-            bucket: "preview",
-            jobId: jobId ?? UUID(),
-            objectPath: "preview/object.jpg",
-            signedUploadURL: "https://example.com/upload",
-            publicObjectURL: nil,
-            expiresInSec: 600
-        )
-    }
-
-    func uploadImageToSignedURL(
-        signedUploadURL: String,
-        contentType: String,
-        imageData: Data
-    ) async throws {
-        _ = signedUploadURL
-        _ = contentType
-        _ = imageData
-    }
-
-    func createNailGenerationJob(
-        shape: NailGenShape,
-        extensionMode: NailGenExtensionMode,
-        handObjectPath: String,
-        referenceObjectPath: String
-    ) async throws -> NailGenCreateJobResponse {
-        _ = shape
-        _ = extensionMode
-        _ = handObjectPath
-        _ = referenceObjectPath
-        return NailGenCreateJobResponse(
-            jobId: UUID(),
-            status: .queued,
-            pollAfterMs: 0
-        )
-    }
-
-    func getNailGenerationJobStatus(jobId: UUID) async throws -> NailGenJobStatusResponse {
-        _ = jobId
-        statusCallCount += 1
-        return statusResponse
     }
 }
