--- NailClient/NailClientTests/FeedChipItemBuilderTests.swift
@@ -1,139 +1,0 @@
-#if false
-//
-//  FeedChipItemBuilderTests.swift
-//  NailClientTests
-//
-
-import Foundation
-import Testing
-@testable import NailClient
-
-@MainActor
-struct FeedChipItemBuilderTests {
-
-    @Test
-    func makeHeaderChipItems_스타일미선택시_카테고리순서유지() {
-        let items = FeedChipItemBuilder.makeHeaderChipItems(
-            categories: ["전체", "스타일", "예약 가능 일정"],
-            selectedCategory: "전체",
-            selectedStyles: [],
-            styleCategoryName: "스타일",
-            scheduleCategoryName: "예약 가능 일정",
-            reservationSummaryText: nil
-        )
-
-        #expect(
-            items == [
-                .category(name: "전체", isSelected: true),
-                .category(name: "스타일", isSelected: false),
-                .category(name: "예약 가능 일정", isSelected: false)
-            ]
-        )
-    }
-
-    @Test
-    func makeHeaderChipItems_스타일선택시_스타일슬롯치환및추가칩노출() {
-        let items = FeedChipItemBuilder.makeHeaderChipItems(
-            categories: ["전체", "스타일", "예약 가능 일정"],
-            selectedCategory: "스타일",
-            selectedStyles: [.natural, .french],
-            styleCategoryName: "스타일",
-            scheduleCategoryName: "예약 가능 일정",
-            reservationSummaryText: nil
-        )
-
-        #expect(
-            items == [
-                .category(name: "전체", isSelected: false),
-                .styleSelected(.natural),
-                .styleSelected(.french),
-                .addStyle,
-                .category(name: "예약 가능 일정", isSelected: false)
-            ]
-        )
-    }
-
-    @Test
-    func makeHeaderChipItems_예약요약있으면_예약카테고리칩을대체한다() {
-        let items = FeedChipItemBuilder.makeHeaderChipItems(
-            categories: ["전체", "스타일", "예약 가능 일정"],
-            selectedCategory: "예약 가능 일정",
-            selectedStyles: [],
-            styleCategoryName: "스타일",
-            scheduleCategoryName: "예약 가능 일정",
-            reservationSummaryText: "2/16 10:00-11:00"
-        )
-
-        #expect(
-            items == [
-                .reservationSummary(text: "2/16 10:00-11:00"),
-                .category(name: "전체", isSelected: false),
-                .category(name: "스타일", isSelected: false)
-            ]
-        )
-    }
-
-    @Test
-    func makeHeaderChipItems_예약요약있으면_예약카테고리칩은노출되지않는다() {
-        let items = FeedChipItemBuilder.makeHeaderChipItems(
-            categories: ["전체", "스타일", "예약 가능 일정"],
-            selectedCategory: "전체",
-            selectedStyles: [.natural],
-            styleCategoryName: "스타일",
-            scheduleCategoryName: "예약 가능 일정",
-            reservationSummaryText: "2/16 10:00-11:00"
-        )
-
-        let hasScheduleCategoryChip = items.contains {
-            if case let .category(name, _) = $0 {
-                return name == "예약 가능 일정"
-            }
-            return false
-        }
-
-        #expect(hasScheduleCategoryChip == false)
-    }
-
-    @Test
-    func makeHeaderChipItems_스타일선택과예약요약동시존재시_순서일관성유지() {
-        let items = FeedChipItemBuilder.makeHeaderChipItems(
-            categories: ["전체", "스타일", "예약 가능 일정"],
-            selectedCategory: "스타일",
-            selectedStyles: [.natural],
-            styleCategoryName: "스타일",
-            scheduleCategoryName: "예약 가능 일정",
-            reservationSummaryText: "2/16 10:00-11:00"
-        )
-
-        #expect(
-            items == [
-                .reservationSummary(text: "2/16 10:00-11:00"),
-                .category(name: "전체", isSelected: false),
-                .styleSelected(.natural),
-                .addStyle
-            ]
-        )
-    }
-
-    @Test
-    func makeHeaderChipItems_예약요약빈문자열이면_예약카테고리칩이노출된다() {
-        let items = FeedChipItemBuilder.makeHeaderChipItems(
-            categories: ["전체", "스타일", "예약 가능 일정"],
-            selectedCategory: "예약 가능 일정",
-            selectedStyles: [],
-            styleCategoryName: "스타일",
-            scheduleCategoryName: "예약 가능 일정",
-            reservationSummaryText: ""
-        )
-
-        #expect(
-            items == [
-                .category(name: "전체", isSelected: false),
-                .category(name: "스타일", isSelected: false),
-                .category(name: "예약 가능 일정", isSelected: true)
-            ]
-        )
-    }
-}
-
-#endif
