+++ NailClient/NailClientTests/TestSupport/Helpers/TestSupportError.swift
@@ -1,0 +1,4 @@
+enum TestSupportError: Error {
+    case forced
+    case unsupported
+}
