+++ NailClient/NailClientUITests/Launch/NailClientUITestsLaunchTests.swift
@@ -30,4 +30,11 @@ final class NailClientUITestsLaunchTests: XCTestCase {
         attachment.lifetime = .keepAlways
         add(attachment)
     }
+
+    @MainActor
+    func testLaunchPerformance() throws {
+        measure(metrics: [XCTApplicationLaunchMetric()]) {
+            XCUIApplication().launch()
+        }
+    }
 }
